build system improvements for verified code
Convert standalone verified/Makefile to a subdir.mk, integrating it
into the build system. It's still possible to build just the verified
code with 'make verified' (or 'make verified/whatever.S'), but this is
the first step to getting the verified monitor linked in place of the
unverified one.

In the process, fixed missing dependencies, and stopped make deleting
common verification DLLs.

Finally, deleted x86test1.