partial post-merge fixes
 * move page_paddr and SecurePhysBaseValue() into smc_handler_spec, avoiding duplication
 * kludge around lack of tuples in spartan by hiding them in a #verbatim block
 * temporarily disable use of mindy (the new smc_handler_spec breaks it)

Unfortunately these are not enough: a number of things don't
verify in Dafny (lots of timeouts, but also some clear problems
that we should look into). It's also not possible to compile the
spec (I get errors about forall statements with no body), so it's
not possible to generate any code.

Next steps: refactoring, {:opaque} where appropriate, and mindy compatibility.