init_l2ptable verifies!
This necessitated some minor refactoring of the relevant pagedb and
smcapi defs, to reuse portions of the spec in the impl proof. I also
backed out the "optimisation" to avoid the early check for a stopped
addrspace... saving a load and conditional branch is not worth the
hours of verification time this was going to take!