Enter verifies!!! Notes: Bitvector math is still handled with assumes in some cases, but there are no other assumes.
On updates to SPSRs ValidModeChange is no longer checked. This should be okay
since we have no instructions that directly change modes based on the contents
of the SPSR. ValidConfig now requires that SPSRS store valid mode encodings
though. If instructions that change modes based on the contents of an SPSR are
added, the mode change should be checked either when the instruction is called
or when the SPSR is modified depending on what the ARM ref actually says should
happen. Previously on SPSR updates the MSR instruction required
ValidModeChange(s, s') where s is the mode in which the instruction is called
and s' is the mode written to the SPSR. This only made sense previously because
only the spsr of the current mode was allowed to be written in the old spec.
Now this would be nonsensical - the spec should actually check that there is a
valid mode change from the mode of the particular SPSR being modified and the
mode being written into the SPSR.

pre_entry_enter now takes over 60s to verify on my machine (despite nothing
changing there except for some asserts to deal with instability). Bumped the
timeout.