Add most of the SMC calls (other than enter/resume) to the top-level smc_handler
This required reworking a good number of the handlers to fix up
static register allocations to match the top-level calling
convention for arguments, and to remove the use of
SysState (which seems irrelevant except for the enter/resume
paths). I also refactored some common code out of init_l2ptable
into a separate file.

Enter/resume are still TODO (due to SysState)

The top-level handler fails to verify (due to missing
postconditions in some of the child handlers), but its
dependencies all verify now.