checkpoint WiP on getting SHA to verify with new decls
Most of the changes here are to work with opaque definitions for
registers, mem and global state validity, which are intended to
decouple memory/global validity from the current state (i.e. once
I have a valid address, it's valid in every state). The problem
with this is a Dafny opaque bug (Dafny issue #63), so there are a
couple of places where we need to reveal the definitions that we
shouldn't have to. There's a lot of noise in this commit, but
most of it is purely mechanical changes from direct map indexing
for mem/globals to use indirect accessors.

Current verification status:
timeouts in sp__refined__Body__00__15, sp__refined__Body__16__XX

errors in sp__abstract__sha256__one__block,
sp__refined__sha256__loop (loop invariants) and
sp__abstract__sha256__block__data__order (register preservation)

I expect most of the errors are due to the opaque bug and hope
that rolling forward to a new Dafny will help.