simplify entry spec to resolve an inconsistency
The new spec was unimplementable, because validExceptionTransition
required that only the dispatcher page changed; this is no longer
true, and in any case there's no good reason for
validExceptionTransition to constrain the pagedb, because it's
computed as a function of svcHandled.

In the process of re-proving things I also simplified
userspaceExecutionAndException which was really only a thin wrapper
over ARMdef's evalUserExecution (and is now explicitly so).